services/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pymongo==3.11.2
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
